Make data objects keep an application context ref
This is necessary for data objects to broadcast events about themselves,
or interact with the application in any other way (though this is rarely
necessary). As coded, it should be possible for data objects to have a
null application context and still behave properly, though their events
will go to /dev/null of course.

This used to be revision r4895.