Implementation of point tool and associated overlays
Note this implementation is not complete. The PointAdapter code still needs to tweak
how points are drawn and how dragging them around works. But as of this commit
one can place points into an image and have them be translated across the legacy
layer.

Note in the future I'll break out a private class that is associated with this commit.

git-svn-id: http://code.imagej.net/svn/imagej/trunk@4899 be23795f-a60e-40d4-ba19-055798872e57