Legacy tests: avoid 'duplicate class definition error'
The problem is that the LegacyInjector wants to redefine ij.ImagePlus. For
that reason, the class loader must not have loaded that class before.
However, if the test class itself defines methods returning an ImagePlus,
they are defined already.

So: move out the helper methods into a standalone class. Since the class
loader must perform lazy loading in method bodies, ImageJ.createContext()
makes sure that the LegacyInjector does its thing before the respective
ImageJ1 classes are loaded.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>

This used to be revision r4915.