Use ij-ext plugin framework for user interfaces
Instead of using SezPoz separately, UserInterface (formerly
IUserInterface) is now defined as a specific type of IPlugin. This
eliminates the requirement for a separate UserInterface annotation
class, and also keeps SezPoz from being a direct dependency of ij-ui.
Instead, it uses the same discovery mechanism as the rest of the ij-ext
framework, which means it will be simpler later to improve the discovery
process beyond "vanilla SezPoz" as needed.

git-svn-id: http://code.imagej.net/svn/imagej/trunk@4922 be23795f-a60e-40d4-ba19-055798872e57