Eliminate ImageJ dependency from native-library-util
As ImageJ is only used to provide logging, it makes sense to eliminate that
dependency in favor of a different logging provider.  For now, we use
java.util.logging.*; this could certainly be changed later as needed.

This change was discussed with Curtis Rueden in OME devteam on 2012-02-27.


This used to be revision r4991.