SwingUI: when a menu bar was added, we need to pack()
It is not enough to invalidate(), but we need to force a repack. Without
that, the scroll view would not have enough space by default and the
scrollbars would be made visible.

Actually, we do not even need to invalidate() the layout explicitly, as
setJMenuBar() does that already for us.

This issue did not occur on MacOSX because the menu bar -- being tucked to
the top of the screen -- is not a part of the window's drawing area there.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>

git-svn-id: http://code.imagej.net/svn/imagej/trunk@4992 be23795f-a60e-40d4-ba19-055798872e57