Expose the ShowPrefs plugin in the File menu
The ShowPrefs plugin was only defined for the mac platform
before this. Linux users could not access the Clear Preferences
functionality. Exposing in the File menu for now. We'll need to
come up with a plan to support application menus for Windows, Mac,
and Linux going forward.

This used to be revision r4995.