pom.xml: use wagon-webdav-jackrabbit
This will allow us to run stuff using Maven 3 (which does not have a
problem with wagon-webdav-jackrabbit but cannot find wagon-webdav).

This change is necessary to accomodate Maven 3, shipped by default with
the 'Eclipse IDE for Java Developers' package.

Suggested by Curtis Rueden instead of a fantastically elaborate
profile-based Maven-only override.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>

This used to be revision r4997.