Use final 1.0 version of wagon-webdav-jackrabbit
This version is the last version to work with both Maven2 (2.2.1) and
Maven3 (3.0.4) from the command line.

Newer versions of wagon-webdav-jackrabbit include 2.0, 2.1 and 2.2, but
all of them fail when running with Maven 2.2.1, with an error message:

  SLF4J: Failed to load class "org.slf4j.impl.StaticLoggerBinder".

Because of this, we will stick to version 1.0 for the time being.

This used to be revision r4998.