Further progress on an About ImageJ2 plugin
Note that this version pulls an AWT dependency into ij-data. This is
very temporary. In DrawingTool there is a class called AWTTextRenderer
that encapsulates all AWT code. The TextRenderer interface and the AWT
implementation will be pulled into UI specific classes. Perhaps there
will be a Graphics or Text service instead.

git-svn-id: http://code.imagej.net/svn/imagej/trunk@4999 be23795f-a60e-40d4-ba19-055798872e57