Update usage of ComputeMinMax class
This adds a dependency on imglib2-algorithms, which is OK since it is now
BSD licensed (the GPL portions of that project were split into a
separate imglib2-algorithms-gpl module).

This change was necessary since the net.imglib2.stats package from
imglib2 was moved to net.imglib2.algorithm.stats in imglib-algorithms.

git-svn-id: http://code.imagej.net/svn/imagej/trunk@5001 be23795f-a60e-40d4-ba19-055798872e57