Adjust JHotDraw figures' line width with the zoom
Rather than setting the width to 0 and drawing the lines ourselves instead
of letting JHotDraw do its thing, convince JHotDraw to do what we want it
to do: to fix the line width at 1 pixel.

Since the figures store the line width in a scale independent manner, we
simply adjust the width according to the current scale whenever the draw()
method is called.

While at it, avoid having private classes when we really want anonymous
ones.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>

This used to be revision r5020.