Add JHotDraw-based ToolDelegator that imitates the IJ1 tool behavior
In ImageJ1, it depends on the location of the mouse cursor whether a new
selection is made or an existing is modified. The ToolDelegator delegates
to the DelegatorSelectionTool or the current IJCreationTool (if a
tool is selected that can create ROIs), depending on the current mouse
cursor position.

It does not move selections yet and does not add to the current selection
when the Shift key is held down.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>

git-svn-id: http://code.imagej.net/svn/imagej/trunk@5026 be23795f-a60e-40d4-ba19-055798872e57