Add script for detecting mixed milestone tickets
A "mixed milestone" ticket is one that is blocked by other tickets from
a different milestone.

Note that this is not always necessarily bad, but it is certainly
undesirable for a ticket from one milestone to be blocked by a ticket of
a later milestone.

This used to be revision r5044.