Begin supporting multichannel assignment of data
With this commit we change how numerous tools and dialogs work.
Rather than dealing with a split between gray and color Datasets
we instead record, track, and draw multiple channels at a time.
Old color data is automatically supported by dealing with 3
channels at a time.

This used to be revision r5051.