Update IJ1 fg/bg colors when picker tool is run
When the IJ2 PickerTool is utilized generate an OptionsEvent. This
makes sure that OptionsSYnchronizer is run. Thus IJ2 picker choices
will update IJ1 colors correctly.

There is some temporary code that accomplishes this. It is clearly
documented in OptionsChannels and in PickerTool. In the future when
we always have a handle on an event service we'll need to always
generate events (via setLastFgColor() and setLastBgColor()) and not
rely on the temp workaround (via setEventService() in PickerTool).

git-svn-id: http://code.imagej.net/svn/imagej/trunk@5058 be23795f-a60e-40d4-ba19-055798872e57