Add AppEventService#getHandledPlugins() method
We now use this method in the MacOSXPlatform to find out which plugins
to remove from the menus, rather than hardcoding the list in two places.

Thanks to Johannes Schindelin for this idea.

This commit also splits the AppEventService interface and default
implementation, so that the platform implmentations have access to the
AppEventService API without adding a dependency on ij-plugins-app (where
the DefaultAppEventService implementation lives).

This used to be revision r5065.