Blacklist some hyperstack related legacy plugins
We will no longer support the legacy behavior of compressing all
dimensions into one slice number slider. Stack to HyperStack and
HyperStack to Stack will be blocked from appearing in the menus.
Users can still run them from macros but the results are not
guaranteed.

This used to be revision r5067.