Minimize time spent in waitForPluginThreads()
waitForPluginThreads() took a very long time to return. I now
realize that it is likely only an issue with the sample image
loading plugins.

I've whitelisted the Image Fetcher threads that stay alive for a
very long time. I believe this is safe. Response time is much more
reasonable now.

git-svn-id: http://code.imagej.net/svn/imagej/trunk@5101 be23795f-a60e-40d4-ba19-055798872e57