Minimize time spent in waitForPluginThreads()
waitForPluginThreads() took a very long time to return. I now
realize that it is likely only an issue with the sample image
loading plugins.

I've whitelisted the Image Fetcher threads that stay alive for a
very long time. I believe this is safe. Response time is much more
reasonable now.

This used to be revision r5101.