Remove obsolete helper class
Sorting plugins by priority is now handled implicitly, due to the fact
that PluginIndex extends SortedObjectIndex, which ensures the lists of
plugins are always sorted. So whenever you ask the index for a list of
plugins matching particular criteria, the list is sorted already.

This used to be revision r5108.