Use plugin mechanism for platform implementations
Rather than having a separate @Platform annotation, Platform is now a
particular type of IPlugin. This transition is similar to the one done
recently for UserInterface, and is part of a general effort to unify
discovery of ImageJ extension points.

This change eliminates the direct dependency on SezPoz for third party
platform authors.

This used to be revision r5118.