Improve text rendering quality
There was a bug in AWTTextRenderer that caused text to get clipped
on the left side sometimes. Now allocating additional buffer space
and drawing into the buffer some units to the right of the buffer
origin. The text rendering code always scans the buffer for nonzero
pixels and determines an origin. So the absolute location within
the buffer is unimportant. Text is no longer clipped.

This used to be revision r5129.