Added support for ImgSaver#isCompressible
The SaveAsImage plugin now takes advantage of ImgSaver#isCompressible to
check its ImgPlus for potential compression requirements (of unknown
Axes of size > 2).

If compression will be required, which necessitates a change in the
ImgPlus structure, the user will be warned and provided a YES/NO dialog
option.

This used to be revision r5189.