Bundle the jar files into jars/ so the ImageJ launcher will pick them up
We will include the ImageJ launcher in a moment, and for
backwards-compatibility this developer thinks it'd be a good thing to keep
the name "jars/" of the directory containing the .jar files.

Since this breaks the precompiled Windows executable, it has been removed for
now, while the Windows batch file, the Shell script and the MacOSX
Info.plist files have been adjusted accordingly.

Unfortunately, the JavaApplicationStub is too limited to parse wildcards
in the classpath, therefore MacOSX continues broken (i.e. this patch does
not fix things).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>

This used to be revision r5194.