Build the ImageJ Launcher for MacOSX Tiger (32-bit), too
The Tiger target is important, as even 64-bit Tigers can only launch
32-bit Java. But a 64-bit Launcher would not even come as far as
determining that it is running on Tiger; dyld would not even let the
program start.

To compile ImageJ-tiger on Leopard, you need to force the profile like
this:

	mvn -Dos.arch=i386 -P 'MacOSX Tiger' -P '!MacOSX Leopard'

(You need to set os.arch, otherwise NAR will mistake the artifact for a
64-bit one.)

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>

This used to be revision r5208.