Bend over for m2e
The 'unpack' goal of the maven-dependency-plugin is apparently not
available via m2e for some strange reason. To shut Eclipse up, we
introduce a profile that is only switched on when not running from
Eclipse (actually, when the 'm2e.version' property is not set, but
that should amount to the same).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>

This used to be revision r5226.