Add ImageJ2 working logo with readme
This image is not the final logo for ImageJ2. The plan is to solicit
community submissions for a logo, then vote on the mailing list to
decide. The current logo is a placeholder image being used during the
beta phase of development.

The ImageJ1 graphics have been moved into the imagej1 subdirectory.

This used to be revision r5236.