Update ImageJ32 and ImageJ64 application bundles
This change removes the unnecessary JavaApplicationStub executables, as
well as the dictionary properties that it used. The ImageJ launcher does
not read such properties, so they serve no purpose anymore.

The ImageJ2 URL is also set to developer.imagej.net, since we are still
in the beta phase and the developer URL is most appropriate for now.

This used to be revision r5242.