Fix profile issue with Swing UI
The problem is that the new "not-eclipse" workaround profile causes the
swing profile's "activeByDefault" to no longer be activated.

We fix the issue by activating all UI profiles based on an "ij.ui"
property. If it is not set, Swing is used; otherwise, the value
indicates the desired UI.

Running "mvn -P all" will still build everything. :-)

This used to be revision r5244.