Eliminate ImageJ32/ImageJ64 dichotomy
We now structure the project more like Fiji, where "ImageJ.app" is the
toplevel folder, so that Mac users get the benefit of a complete
application bundle.

Both 32-bit and 64-bit ImageJ launchers are present anyway, so there is
no need for two separate icons.

This used to be revision r5248.