ImageJ launcher (windows): be nice when falling back to system Java
Most users do not really care which Java was used, so do not open a
console (and consequently fall back to java.exe instead of javaw.exe,
opening another console) just to say we fell back to launching
system Java.

Noticed by Jimmy Fong.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>

This used to be revision r5268.