Updater: stageForUpload() marks uninstalled files for removal now
The 'upload' button uploads new files and an updated db.xml.gz. In
that sense, when a user marks a not-installed file for upload, it is
logical to assume that db.xml.gz needs to be uploaded after marking
the not-installed file as obsolete.

Thanks Jody Jenkins.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>

This used to be revision r5269.