Do not rely on up-to-date Apple Java Extensions on MacOSX
On MacOSX Tiger, the Update 8 of Java might not be installed, and even on
MacOSX Leopard, Update 3 of Java might not be installed yet. But according
to

	http://developer.apple.com/library/mac/documentation/Java/Reference/JavaSE6_AppleExtensionsRef/api/com/apple/eawt/AboutHandler.html

the AboutHandler (and all the other interfaces MacOSXAppEventDispatcher
implements) is available only starting with Update 8 for Java on MacOSX
10.5 and with Update 3 for Java on MacOSX 10.6.

Let's err gently if these updates have not been applied yet. In
particular, let's not let the platform service prevent the UI service from
instantiating properly (symptom: ImageJ window opens briefly but closes
right away).

Noticed by Kevin Eliceiri.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>

This used to be revision r5279.