Add hack for ij-launcher project in Eclipse
This is similar to the hack for ij-app, except that we must also hack
the packaging type to be "jar" instead of "nar" when working in Eclipse.
This must be done through a property because you cannot set the
packaging directly within a profile.

It is an ugly hack but does result in the needed packaging type from
both command-line Maven and Eclipse.

This used to be revision r5303.