Fix IJ1 classpath problem with third-party jar libraries
When a IJ1 plugin depends on other libraries in the jars folder, the
dependent classes cannot be loaded.

For example, add the LOCI update site ( http://dev.loci.wisc.edu/fiji/),
install the plugins, then run the Spectral Lifetime Analysis plugin. An
error is emitted:

	Plugin or class not found: "slim.SLIM_PlugIn"

This happens because a class in
jars/curve-fitter-1.0-SNAPSHOT-with-libs.jar cannot be found, and then
ImageJ1 decides to get creative.

This issue was reported by Curtis Rueden as ticket #1094.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>

This used to be revision r5305.