Updater: add methods to get at the filename without version
We'd like to be able to determine the unversioned file name of .jar files
(e.g. bio-formats.jar for bio-formats-4.4.jar) to be able to determine
whether two .jar files are just two different versions of the same
artifact.

This is important for the code allowing to upgrade to files with a higher
version number, as required by ImageJ2 where all the files keep the
version in their filename.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>

This used to be revision r5365.