LegacyPlugin - when waiting ignore timer threads
The waitForPluginThreads() code hatches a timer to accomplish its
work. Then the wait loop can detect the thread as not terminated
and wait forever. Fixed.

Now whitelist(thread) accepts all "Timer*" threads. This should be
okay as a timer threads are associated with suspended worker threads.
It is the worker threads we want to test for completion.

This used to be revision r5426.