ImageJ launcher: include client/ or server/ in LD_LIBRARY_PATH
When trying to dlopen libjvm.so, we need to find libjsig.so, too, which
is stuck inside /lib/i386/client/ in case of 32-bit Linux. So let's tuck
that path onto LD_LIBRARY_PATH, too, in case of old Java which does not
handle LD_LIBRARY_PATH issues gracefully.

While at it, output the full LD_LIBRARY_PATH on the command-line.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>

This used to be revision r5428.