Streamline AWT input event dispatching logic
This eliminates the separate mouse and key event dispatchers, in favor
of a unified dispatcher. (AWTWindowEventDispatcher is still separate.)

A central reason for this unification is to track last known mouse
coordinates, so that KyEvents can pass along that information
painlessly. This is useful for things like zooming in and out with the
+/- keys, where it would be nice to zoom at the current mouse
coordinates rather than only the center of the viewport. ImageJ1 uses
the same trick, tracking last known mouse coordinates using a mouse
listener, and then using those values as needed in response to non-mouse
events as well.

This used to be revision r5443.