Add Display.isDisplaying(Object) method
This is a general method similar to contains(Object) from the
collections API, except that it is more flexible. In particular, in the
case of an ImageDisplay, it will work as expected when asking about a
particular Data object. Since ImageDisplay actually displays
DataView objects, not Data objects, ImageDisplay.contains(Object) will
always return false, and rightly so according to the collections API.

This change allows the DisplayPostprocessor to properly query whether
outputs of type Data are currently being displayed; previously, a new
display would always be created.

Fixes #1157.

This used to be revision r5456.