Updater: store then-current file names with the previous versions
If we every want to be able to revert to previous states with the
updater, we have to make sure that we can figure out the correct file
names on the update site...

If, say, ij-updater-core-Vista.jar is current, but we want to go
back to a state where the file was called ij-updater-core-NT.jar,
that information has to be available.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>

This used to be revision r5462.