Updater: be nice to old-style update sites
There are update sites -- notably the Fiji update site -- which
were set up with the idea that files' contents can change but never
the name.

With the introduction of the version in the file names (which we want &
need in ImageJ2), that notion changed.

As a consequence, the existing db.xml.gz files can contain multiple
entries that would now become one. Handle that situation by making
them one.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>

This used to be revision r5466.