Updater: initialize metadata from pom.xml
When (Mini-)Maven generates a .jar file, it is nice enough to put the
pom.xml into a subdirectory of META-INF/. We can gleen the description
and the authors from it directly, and we may well be able to also
determine the dependencies in the future. After defining a standard,
we may even be able to put information about links, platforms and
categories there.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>

This used to be revision r5468.