Updater: fix overall progress on upload
We were a bit overzealous trying to kill the progress bar's performance
issues. While it is true that the progress updates in the checksummer
slowed it down by an order of magnitude, we should make a difference
between the updates to the overall progress and the individual items'
progress bar. The uploader exposed that issue, as it updated the
item-specific progress bar first, then the overall one, which did
not update at all as a consequence.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>