Add maven-java-formatter-plugin configuration
This uses the Eclipse formatter to do the work, but wraps it in an
easy-to-use Maven plugin.

You can run it using:
  mvn java-formatter:format

More details at:
  http://code.google.com/p/maven-java-formatter-plugin/

Unfortunately, similar to other uses of the Eclipse formatter from the
console, it only runs the Format command, and not the additional
features of the "Clean Up..." command. But it is better than nothing!

This used to be revision r5473.