Do not persist objects that cannot be restored
Such objects include Dataset and Display.

In the future we may want a more sophisticated mechanism for controlling
how an object is "restored" from a persisted string. For example,
Dataset expresses itself as a string, which could later be matched to
the name of an existing Dataset. But currently we have no facility for
that, so for now it is best simply to avoid persisting such information.

This used to be revision r5481.