Remove unnecessary required and persist flags
The required flag is true by default, and the persist flag is no longer
necessary on objects that cannot be restored from a persisted string,
since the default persistence mechanism now checks for that in advance
and avoids persisting objects it cannot restore later.

This used to be revision r5482.