Fix menu bar creation & associated event handling
The Swing UI *always* needs to subscribe to events, regardless of
whether isMenuBarDuplicated returns true. And the menu bar should only
be added to displays if isMenuBarDuplicated returns true. (This was
previously be the case, but with the display refactoring, it got lost in
the shuffle.)

This used to be revision r5518.