Fix bug that could cause long delays
When running legacy plugins simultaneously the waitForPluginThreads()
code could fail to terminate for a considerable amount of time. Now
LegacyPlugin::run() is updated to launch a new Thread that handles
running the plugin and avoids contention between different plugins.
The contention is avoided by inspecting which threadgroup a blocking
thread lives in. A plugin only waits for threads it has hatched to
terminate.