Fix oversight in OverlayFigureView
The figureRemoved() method was passing the wrong kind of object to
display.remove(o). This makes things more consistent but further breaks
zooming.

Before this change zoom was being reset upon overlay creation. Now it
resets zoom on overlay creation and overlay deletion.

To fix these we need to track structure changes in a different way for
AbstractDisplays (and Displays in general).