Check super-types for annotated parameters first
This commit results in clearer error messages when fields with
conflicting names are discovered. A superclass field Foo#data with a
subclass field Bar#data will now report that Bar#data is the erroneous
duplicate, rather than Foo#data.

In the case of plugins with parameters defined in a class hierarchy,
this commit changes the canonical order of the parameters. This will
affect the ordering in the input harvester dialog. It is also important
when calling e.g. PluginService#run(String, Object...), as the order of
the passed-in values must match the canonical parameter ordering.