Add the script to update glibc-compat.h
This script was developed during the time when the launcher was still
known as the Fiji launcher. Its purpose is to update the header forcing
the symbol versions so we're staying bugwarts-compatible: GNU libc
versions contain these versioned symbols which I am sure solves some
problems in addition to the many problems it causes. The symptom of the
problems is something like

	/lib64/libc.so.6: version `GLIBC_2.14` not found

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>