ImageJ launcher (Linux): fix LD_LIBRARY_PATH logic
Ilan Tal noticed that we do not adjust the LD_LIBRARY_PATH correctly, or
at least that we do not always re-execute Fiji when necessary.

The problem arises when you have a library depend on another library which
is in a different directory. Example:

	$IJDIR/lib/linux64/libJNI_Example.so

depends on

	$IJDIR/lib/misc/libCoolStuff.so

Both are in the java.library.path, but that does not matter because we
only load libJNI_Example and the dynamic linker (ld.so) is happily
oblivious of java.library.path. It needs LD_LIBRARY_PATH to be set
correctly, which means _before_ our code is executed (i.e. we need to
re-execute).

Now, Java6 still has no problem to find its libraries with LD_LIBRARY_PATH
left alone, but if we have other libraries, we need to re-execute so that
the LD_LIBRARY_PATH is correct, even for Java6.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>