Sezpoz checker: verify that annotations really changed
If the annotation processing runs but does not produce different
annotations, we should pretend that the annotations were not really
processed.

That way, we can ask the Sezpoz checker: can we proceed now? Or does the
user have to restart Java so that the annotations are properly discovered
by the ClassLoader's getResources() method?

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>