Allow platforms to override behavior of JMenuBar
After creating the application menu structure, the Swing UI broadcasts
an AppMenusCreatedEvent informing the world. The PlatformService picks
it up, and gives its active platform handlers a chance, one at a time,
to do some platform-specific handling of the newly created menus via the
Platform.registerAppMenus(Object) method. If one of them exercises this
option, the AppMenusCreatedEvent is consumed to flag it, and the Swing
UI declines to add the JMenuBar to the SwingApplicationFrame.

This allows the MacOSXPlatform to perform special handling of the menu
structure, while leaving the default behavior (for Windows, Linux, etc.)
as a JMenuBar attached to the main application frame.

This commit should fix bugs surrounding the application menu bar on both
Mac and non-Mac systems.