Improve the event service implementatation
Previously, when publishing an event on the EDT, the publication order
would resemble a stack because the publication would happen immediately,
but when publishing an event off the EDT, the EDT would be used via
EventQueue.invokeLater, which would cause the publication order to
resemble a queue instead. This inconsistency made it difficult to
predict and debug behavior.

We now explicitly provide a publish method (which publishes immediately,
blocking until publication is complete), as well as a publishLater
method (which queues the event for publication on the EDT at some future
date). The gives the publishing code control over how its event is
consumed; if it needs to know how subscribers reacted to continue, it
can call publish to block until publication is fully completed.

To accomplish this, new methods have been added to the ThreadService
interface for executing code on the EDT immediately (invoke), as well as
queuing code for later execution on the EDT (queue). And the
EventService now depends on the ThreadService to access this
functionality.