Fix blurry text on Retina displays
When testing ImageJ 1.* on the new Retina displays on the new MacBook
Pros, text, graphics and controls look blurry, and the "Open in Low
Resolution" option in the Get Info dialog is checked and can't be
unchecked.

The trick for getting a Java application to have high-resolution text,
graphics and controls on a MacBook Pro with Retina display is making
sure it has a Principal Class setting of NSApplication in the Info.plist
file.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>