Execute preprocessors on OptionsPlugins
The OptionsService was instantiating OptionsPlugin instances, but did
not run the preprocessor chain on them, so any @Parameters they had were
not autopopulated, any initializers were not run, etc.

Now, the preprocessing occurs, so they should be fully initialized.

This fixes a bug in OptionsChannels that was previously worked around by
manually passing the EventService instance from PickerTool.