Improve signatures of some PluginService methods
Specifically, the getPluginsOfType, createInstancesOfType and
createInstances methods used a generic parameter P (P extends IPlugin).
They returned List<P> and/or List<PluginInfo<P>>, when it would be more
appropriate and flexible to use List<? extends P> and/or
List<PluginInfo<? extends P>>.

The ramifications of this change propagate through the system in a few
places, but the changes are worth it as they make it much easier to
invoke PluginService.createInstances on a mixed list of PluginInfos.