Move UncaughtExceptionHandler logic to LogService
Our default uncaught exception handler uses the log anyway, so it makes
sense to provide it with a LogService to use. Then the DefaultLogService
can take care of registering itself (essentially) as the default
uncaught exception handler.

The only problem is that because Java threads are shared across all
active application contexts, it is difficult to determine which
application context was involved with a particular uncaught exception.
Right now, every new DefaultLogService simply registers itself as *the*
default uncaught exception handler, which is a bit strange when there
are multiple simultaneously active application contexts. But I don't
have a better idea right now how to deal with it.