Have ShadowMenu track an application context
Previously, ShadowMenu tracked a MenuService, and from that it derived
access to certain other services (specifically, EventService,
PluginService and LogService). However, it is simpler to have it linked
an application context directly, and access relevant services from that.

Note, however, that when writing code this way, there is no hard
guarantee that the given context will have any particular services. It
is best to write all code as though the requested services may not be
present -- so you'll notice that ShadowMenu now tests for null regarding
all services it uses. This is actually nice for ShadowMenuTest, which
can now pass an empty context and still verify the behavior just fine.