A better WatchEvents implementation
This version uses the EventHistory service to track things.

It displays a tree on the left showing the event hierarchy of all events
that have occurred since monitoring started.

On the right, it shows an HTML pane with the event feed, syntax
highlighted for easier reading. Selecting nodes in the tree will display
those types of events in bold in the event feed.