LegacyService: ignore overridden legacy plugins
We have a number of pure-IJ2 plugins that occupy the same locations in
the menu structure as their overridden IJ1 counterparts (which are then
obsolete). There are various reasons for reimplementing such plugins;
often it is either:

  A. to improve the plugin to utilize new features of IJ2; or
  B. because the legacy plugin does not work in the IJ2 framework.

Previously, we would manually add such overridden legacy plugins to the
plugin-blacklist.txt file, which would prevent them from being added to
the plugin index. Unfortunately, this is not an extensible approach.

Hence, we have now added a special case in the LegacyService to ignore
any legacy plugin whose location in the menu structure is already taken
(presumably by a pure-IJ2 plugin equivalent). It is still possible to
manually blacklist legacy plugins that do not fit this paradigm, but the
list should become substantially shorter now.

Thanks to Johannes Schindelin for the suggestion!