Refactor management of display viewers
The logic was previously Swing-centric, existing mostly in
AbstractSwingUI, but the logic itself was almost completely
Swing-ignorant (with the exception of the DisplayWindow creation).
There was also some duplicate logic in SwingUI and SwingMdiUI.

This commit factors out the logic into the UI service implementation:
	* The UI service itself (rather than the Swing UI) now tracks which
    display viewers exist, and provides a getter for accessing viewers.
  * The UserInterface interface now provides a means to create a new
    DisplayWindow for a given Display.
  * The AWTWindowEventDispatcher and InternalFrameEventDispatcher now
    have a simpler constructor sans EventService argument, since that
    service can be extracted from the given Display's context.

All in all, this refactoring resulted in a net reduction in code volume!