Add label for app plugins
The reason is that on Mac OS X, if they are removed from the menu
structure, they lose their menuPath, including the label of the menu
leaf. So the Command Finder falls back to calling them by their class
name, which is sort of ugly. By giving them a proper plugin label, that
name is preferred on all platforms for things like the Command Finder.