Make a ImageCapture plugin
Used by (and adapted from) SwingOverlayManager::flatten(). A number of
dependencies changed with this. Also UI's can now be queried for their
ImageDisplayViewer associated with an ImageDisplay (if it exists). This
is a way to minimize the changes necessary to the Display hierarchy.

This plugin replaces IJ1's Flatten command in the menus too.