Make AWTInputEventDispatcher adjust for pan offset
This fixes a problem where the mouse coordinates passed through to the
MsEvents do not match up with the panel coordinates of the ImageCanvas.

The ImageCanvas expects (0, 0) to be the top left corner of the active
viewport (i.e., regardless of the pan center). But the
JHotDrawImageCanvas must add a listener to its DrawingView which is
wrapped in a JScrollPane, resulting in coordinates offset from the top
left corner of the DrawingView. Hence, when the scroll pane is not at
the top left position, the DrawingView coordinates do not match the
viewport's coordinates, but are instead offset by an amount matching the
pan offset.

To address this, we check whether the event's linked Display is an
ImageDisplay, and if so, obtain that ImageDisplay's associated
ImageCanvas, then compute the pan offset and subtract from (X, Y) before
passing the coordinates on to the newly minted MsEvents.

Future AWT-based implementations (notably in the as-yet-unfinished
pure-AWT UI) are expected to work properly, as long as the Component to
which the AWTInputEventDispatcher is attached is also wrapped in a
ScrollPane or similar viewport component that manages the pan similarly
to how JScrollPane works. However, if this assumption does not hold, we
may need to add a boolean flag to the AWTInputEventDispatcher
constructor to allow explicit toggling of the pan adjustment behavior.