JHotDrawImageCanvas: add threading sanity check
We want to make sure we are always on the dispatch thread (in the case
of AWT/Swing, this will specifically be the AWT Event Dispatch Thread).

Once we know that we are on the EDT, we don't have to worry about
multi-threaded race conditions, or calls to pack() causing any trouble.