Image canvas: use better zoom center approach
When specifying a point at which to zoom (in or out), it makes intuitive
sense for that point of focus to remain stationary, while everything
else zooms around it. This allows for repeated zooming in and out at the
same point, such as with the zoom tool.

Thanks to Johannes Schindelin for pointing it out.