Remove unnecessary EventQueue & synchronized code
The entire chain of events set off by ImageJ's event bus, and/or by
changes in the UI such as those of JHotDraw, happens on the EDT. So it
is completely unnecessary, and in fact may introduce errors in the order
of operations (due to invokeLater queuing things rather than executing
them immediately), to use EventQueue.invokeLater.

There are some public methods which are now unprotected by synchronized
and/or invokeLater. I think this is OK, and is should be the
responsibility of downstream code, when in doubt, to wrap calls in
ThreadService.invoke and/or ThreadService.queue (which both delegate to
EventQueue in the default implementation).

In general, I think removing the synchronized keyword is good, both for
performance, and also to avoid potential deadlocks -- though if any
deadlock were possible here, it would be due to multiple threads calling
these methods simultaneously, which should not be the case currently.