Fix overlay line width behavior
However, this change removes the logic that scales the line width based
on zoom factor. To maintain fixed-width lines regardless of zoom, we
should readd that feature. Proposed strategy for doing that:

  1. Change the line width synchronization logic in
     AbstractJHotDrawOverlayAdapter#updateFigure to scale the JHotDraw
     Figure's STROKE_WIDTH property inversely to zoom factor.
  2. Call updateFigure in response to PanZoomEvents.

The latter could be done by the OverlayService and would not require any
case logic in any of the AbstractJHotDrawOverlayAdapter subclasses.