Fix opacity bug in overlay synchronization
We now always use a plain ColorRGB (not a ColorRGBA) for the Overlay's
fillColor. The alpha is stored in a separate int parameter called alpha.
Previously, the fillColor was mixing in the alpha parameter, which
resulted in a painting bug in the Overlay Properties dialog, where the
fill color would paint more darkly due to the decreased alpha value.