Updater: re-do the obsolete checksum handling
The earlier patch to handle update sites with db.xml.gz files containing
obsolete checksums gracefully was flawed. For one, we should not pretend
to know better than the update site, and additionally, the way we solved
the issue did not allow for caching the checksums easily.

Instead, let's just pretend that our locally-calculated checksum actually
matched the one from the update site.

For this to work, we need to store not only the current checksum but also
calculate the obsolete checksums and cache them, too.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>