Updater: add a convenience function
In Fiji, we would like to have an easy way to install the ImageJ Updater.
To do so, we will use a then-current version of the Updater via a
URLClassLoader accessing the ImageJ Update site to install itself. After
that, we need to make a new classloader to access the newly-installed
classes, hence the need for a list of dependencies, including transitive
ones.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>