ImageJ launcher: be quiet when Fiji's IJ1 patcher was not found
As a rule, the ImageJ launcher -- as it used to be the Fiji launcher
-- wants to patch ImageJ 1.x in a few ways to be able to a few cool
things such as using Fiji's Script Editor in place of the plain text
editor e.g. when clicking the 'Edit' button in the Recorder.

However, when we do not even have Fiji's IJ1 patcher class, it is
quite safe to assume that we are not running inside Fiji anyway...

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>