Updater GUI: offer to move the files into place without restart
This is not the safest thing to do, as Java relies on the .jar files
being available in unchanged form. This was done so that older Windows
versions -- which were seriously limited in file descriptors --
would not run out of resources just because the class path contains
593 .jar files.

As a consequence, updating .jar files while they are in use can lead to
'funny' exceptions and even crashes.

Nevertheless, for people who know what they're doing (or think they
do), having the option of shooting yourself into your foot is nicer
than not to have it.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>