Add an OutputStream adapter for line-based outputs
When parsing the output of programs, it is often convenient to have
an adapter that passes complete lines instead of random chunks of
bytes. This class helps with that situation. Its initial purpose is
to help with colorizing the differences between local and remote .jar
files in the ImageJ updater.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>