Updater GUI: offer to show how local and remote versions differ
This is a major overhaul of the 'Show changes' button in the advanced
mode. It also switches this mode on based on the availability of the
'git' executable rather than the presence of the .git/ directory in
the ImageJ root directory, since the latter is no longer the case
for ImageJ2.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>