Updater GUI: Really make sure that the context class loader is set correctly
This solves the problem that threads launched from the EDT did not have
the same context class loader as the the thread that instantiated the
Updater GUI.

It is a more complete solution than ed6fd29a(Updater GUI:
Make sure that the context class loader is set correctly): that worked
only for uploading, but not for initializing a new update site.

Noticed by Olivier Burri.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>