InputWidget: add new API methods
The setModel(WidgetModel) method allows the WidgetModel to be assigned
after construction.

The isCompatible(WidgetModel) method checks whether a given widget is
appropriate for use with a particular widget model object. For example,
a number slider widget would be appropriate for use with a WidgetModel
housing a ModuleItem of type Number, and NUMBER_SLIDER WidgetStyle.

These methods will be important later for the extensible widget
mechanism, so that we can construct initially "empty" widgets and then
test them for compatibility with each module input to be harvested.