Merge ij-platform-macosx and ij-platform-windows
They have become a single module called ij-platforms now with base
package imagej.core.platforms, similar to core/plugins and core/tools,
which dwell in imagej.core.plugins and imagej.core.tools respectively.

The rationale for this change is to reduce the total number of modules,
to make the system easier to understand. The PlatformService is designed
to trigger applicable platforms only, so having both WindowsPlatform and
MacOSXPlatform in the classpath is not harmful at all; in this way,
developers can choose to bundle ImageJ's default desktop platform
handling logic merely by including a single ij-platforms JAR.